Corrected problem with getTotal() in Customer.

+++ ci/jenkins/javaAntExample/src/uk/ac/aber/agile/movies/Customer.java
@@ -75,7 +75,7 @@ public class Customer {
 	}
     
     private double getTotalAmount() {
-		int amount = 0; 
+		double amount = 0; 
 		for(Rental rental : customerRentals) { 
 		   amount += rental.getCharge();
 		}
